batch:p1 convert batchJob to new Batch template
